Ensure direct tool requires conflicts are properly reported and Conan does not hang
---------
